Allow output to be written to a file
The behavior of the programs remains the same when the program is called as
this:

./ps1_reverse <PS-X EXE>

But now the user has the option of writing the output to a file:

./ps1_reverse <PS-X EXE> [OUTPUT FILE]

In this case, nothing is written to the std::cout, which helps to improve the
performance. Also, visualizing the output of a big EXE in a file is much
easier, as it is the case of a typical PS1 game.